+++ components/current-position.vue
@@ -1,12 +1,13 @@
 <template>
   <div class="fixed-pitch">
-    {{ position }}
+    {{ position.substring(0, 8) }}
+    <span class="millisec">{{ position?.substring(8) }}</span>
   </div>
 </template>
 <script setup lang="ts">
-const invalidPosition = '00:00:00.000'
 let animationFrameId = 0
 
+const invalidPosition = '00:00:00.000'
 const position = ref(invalidPosition)
 
 const update = () => {
@@ -18,8 +19,8 @@ const update = () => {
     const hours = String(date.getUTCHours()).padStart(2, '0')
     const minutes = String(date.getUTCMinutes()).padStart(2, '0')
     const seconds = String(date.getUTCSeconds()).padStart(2, '0')
-    const milliSeconds = String(date.getUTCMilliseconds()).padStart(3, '0')
-    position.value = `${hours}:${minutes}:${seconds}.${milliSeconds}`
+    const milliSeconds = date.getUTCMilliseconds()
+    position.value = `${hours}:${minutes}:${seconds}${milliSeconds >= 500 ? '' : '.'}${String(milliSeconds).padStart(3, '0')}`
   }
   animationFrameId = requestAnimationFrame(update)
 }
@@ -38,7 +39,11 @@ useFbdPlayer().addEventListener((ev) => {
 </script>
 
 <style type="scss" scoped>
-.fixed-patch {
-  font-family: "Fira Mono", ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
+.fixed-pitch {
+  font-family: "DSEG7-Modern", "Fira Mono", ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
+  font-size: 0.75rem;
+}
+.millisec {
+  font-size: 0.6rem;
 }
 </style>