+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.10] - 2024-08-07
+
+### Added
+- Added DSEG7 font.
+- Added display of peak hold value.
+
+### Fixed
+- Correct CSS class name errors.
+
 ## [0.1.9] - 2024-08-06
 
 ### Changed